Add padding to FillLayout
Also made Layout a reactive object but I chose not to dispose it with the
parent component because it'd violate the "Boy Scout Rule".

At least, any reference it may contain should be confined to itself and
maybe the parent component, so I suppose it wouldn't cause a serious memory
leak if one does not share a layout between components (which you shouldn't).